+++ src/modules/hatcher/registries.js
